requirements.txt
art==5.1
asgiref==3.4.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
cffi==1.14.4
distlib==0.3.3
Django==3.2.13
django-autocomplete-light==3.8.1
django-auth-ldap==2.2.0
django-braces==1.15.0
django-crispy-forms==1.13.0
django-decorator-include==3.0
django-environ==0.8.1
django-extensions==3.1.5
filelock==3.3.2
graphviz==0.19
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.20
gunicorn==20.1.0
pydot==1.4.2
pyparsing==3.0.6
pytz==2021.3
django-bootstrap-icons==0.7.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
django-bootstrap-datepicker-plus==4.0.0
django-phonenumber-field==2.1.0
django-bootstrap-v5==1.0.11 
django-bootstrap-form==3.4
typing-extensions==4.0.1
whitenoise==5.3.0
django-filter==2.4.0